bug(app): app wide bug fixes and refactoring
- fix bugs on different files
- modify database files to use ES6
- install morgan to log requests to the server
- install debug to replace console.log
- some minor bug fixes

[Delivers #166358637]
