+++ .eslintrc.json
@@ -4,7 +4,7 @@
   "env": {
     "node": true,
     "es6": true,
-    "mocha": true
+    "jest": true
   },
   "rules": {
     "one-var": 0,
@@ -13,6 +13,7 @@
     "consistent-return": 0,
     "no-param-reassign": 0,
     "comma-dangle": 0,
+    "no-console": 2,
     "curly": ["error", "multi-line"],
     "import/no-unresolved": [2, { "commonjs": true }],
     "no-shadow": ["error", { "allow": ["req", "res", "err"] }],
