+++ package.json
@@ -7,7 +7,7 @@
     "start": "node build/index.js",
     "test": "jest --forceExit --verbose",
     "coverage": "jest --forceExit --coverage && cat ./coverage/lcov.info | coveralls",
-    "dev": "nodemon --exec babel-node server/index.js",
+    "dev": "cross-env DEBUG=dev && nodemon --exec babel-node server/index.js",
     "clean": "rm -rf build && mkdir build",
     "build-server": "babel -d ./build ./server",
     "build": "npm run clean && npm run build-server",
@@ -34,8 +34,12 @@
   },
   "homepage": "https://github.com/emp-daisy/HelloBooksLib#readme",
   "dependencies": {
+    "body-parser": "^1.19.0",
+    "cors": "^2.8.5",
+    "debug": "^4.1.1",
     "dotenv": "^8.0.0",
     "express": "^4.17.1",
+    "morgan": "^1.9.1",
     "pg": "^7.11.0",
     "pg-hstore": "^2.3.2",
     "sequelize": "^5.8.6",
@@ -50,6 +54,7 @@
     "@babel/preset-env": "^7.4.5",
     "@babel/register": "^7.4.4",
     "coveralls": "^3.0.3",
+    "cross-env": "^5.2.0",
     "dotenv": "^8.0.0",
     "eslint": "^5.16.0",
     "eslint-config-airbnb-base": "^13.1.0",
