+++ server/db/models/index.js
@@ -1,11 +1,11 @@
-'use strict';
+/* eslint-disable import/no-dynamic-require */
+import { readdirSync } from 'fs';
+import { basename as _basename, join } from 'path';
+import Sequelize from 'sequelize';
 
-const fs = require('fs');
-const path = require('path');
-const Sequelize = require('sequelize');
-const basename = path.basename(__filename);
+const basename = _basename(__filename);
 const env = process.env.NODE_ENV || 'development';
-const config = require(__dirname + '/../config/config.json')[env];
+const config = require(`${__dirname  }/../config/config.js`)[env];
 const db = {};
 
 let sequelize;
@@ -15,13 +15,12 @@ if (config.use_env_variable) {
   sequelize = new Sequelize(config.database, config.username, config.password, config);
 }
 
-fs
-  .readdirSync(__dirname)
+readdirSync(__dirname)
   .filter(file => {
     return (file.indexOf('.') !== 0) && (file !== basename) && (file.slice(-3) === '.js');
   })
   .forEach(file => {
-    const model = sequelize['import'](path.join(__dirname, file));
+    const model = sequelize.import(join(__dirname, file));
     db[model.name] = model;
   });
 
@@ -34,4 +33,4 @@ Object.keys(db).forEach(modelName => {
 db.sequelize = sequelize;
 db.Sequelize = Sequelize;
 
-module.exports = db;
+export default db;
